refactor: migrate service to providers
