+++ src/hooks/api.tsx
@@ -1,4 +1,4 @@
 import { useContext } from 'react';
-import { ApiContext, ApiCtx } from '../service';
+import { ApiContext, ApiCtx } from '../providers';
 
 export const useApi = () => useContext(ApiContext) as Exclude<ApiCtx, null>;
