+++ src/hooks/assets.tsx
@@ -1,4 +1,4 @@
 import { useContext } from 'react';
-import { AssetsContext, AssetsCtx } from '../service/assets-provider';
+import { AssetsContext, AssetsCtx } from '../providers/assets-provider';
 
 export const useAssets = () => useContext(AssetsContext) as Exclude<AssetsCtx, null>;
