+++ src/providers/assets-provider.tsx
